+++ examples/typescript/CustomTelemetryHandler.ts
@@ -1,0 +1,8 @@
+import type { TelemetryHandler } from '@shellicar/winston-azure-application-insights';
+import type { TelemetryData } from '../../src/public/types';
+
+export class CustomTelemetryHandler implements TelemetryHandler {
+  handleTelemetry(telemetry: TelemetryData) {
+    console.log('Custom Telemetry Handler:', telemetry);
+  }
+}