+++ examples/typescript/main.ts
@@ -1,13 +1,13 @@
 import { createApplicationInsightsTransport } from '@shellicar/winston-azure-application-insights';
-import { defaultClient, setup } from 'applicationinsights';
+import applicationinsights from 'applicationinsights';
 import { createLogger, format, transports } from 'winston';
 
 // By default uses process.env.APPLICATIONINSIGHTS_CONNECTION_STRING
-setup().start();
+applicationinsights.setup().start();
 
 const transport = createApplicationInsightsTransport({
   version: 3,
-  client: defaultClient,
+  client: applicationinsights.defaultClient,
 });
 
 const logger = createLogger({